+++ test/dictionary/blank.wav
@@ -1,0 +1,1 @@
+NOT A VALID WAV FILE
